fix: jaeger Tracing 코드 정리
